requirements.txt
blinker==1.6.2
click==8.1.6
dnspython==2.6.1
Flask==3.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pymongo==4.8.0
python-dotenv==1.0.1
Werkzeug==3.0.0
